requirements.txt
streamlit 
tensorflow 
pillow 
matplotlib
streamlit-webrtc
opencv-python-headless
scikit-learn
pandas
plotly
seaborn